requirements.txt
aiohttp==3.9.0
aiogram==3.20.0
aiogram-dialog==2.2.0
celery==5.3.6
Django==5.1.4
django-celery-beat==2.8.0
django-cors-headers==4.3.1
djangorestframework==3.16.0
gunicorn==22.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
redis==5.0.3
